requirements.txt
alembic==1.7.7
Flask==2.0.1
Flask_Bcrypt==0.7.1
Flask_Cors==3.0.10
Flask_JWT_Extended==4.3.1
flask_marshmallow==0.14.0
Flask_Migrate==4.0.0
Flask_SQLAlchemy==2.5.1
marshmallow_sqlalchemy==0.23.1
SQLAlchemy==1.3.19
gunicorn==19.9.0